# shared/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateTripRequest": {
        "description": "Create a trip. Everything is public and unauthenticated for now.",
        "properties": {
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Explicit slug. Omit to derive one from the name.",
            "title": "Slug"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateTripRequest",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Uniform error body. `code` is stable and machine-readable; `detail` is for humans.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "code",
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "PoiCategory": {
        "description": "What a point of interest is, which drives its map icon and discovery prompt.",
        "enum": [
          "wild_camp",
          "campground",
          "hotel",
          "unique_stay",
          "food",
          "fuel",
          "water",
          "viewpoint",
          "mechanic"
        ],
        "title": "PoiCategory",
        "type": "string"
      },
...
      "ReplanEvent": {
        "description": "One streamed progress event from a replan.",
        "properties": {
          "legs": {
            "description": "Legs re-routed by this stage, if any.",
            "items": {
              "$ref": "#/components/schemas/TripLeg"
            },
            "title": "Legs",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "pois": {
            "description": "POIs discovered so far in this stage. Cumulative per stage, not total.",
            "items": {
              "$ref": "#/components/schemas/Poi"
            },
            "title": "Pois",
            "type": "array"
          },
          "progress": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "stage": {
            "description": "One of: route_search, discovery, enrichment, done.",
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "stage",
          "message"
        ],
        "title": "ReplanEvent",
        "type": "object"
      },
      "ReplanRequest": {
        "description": "Start the slow path. Explicitly user-triggered; never fired by a route edit.",
        "properties": {
          "categories": {
            "description": "Which kinds of POI to discover along the route.",
            "items": {
              "$ref": "#/components/schemas/PoiCategory"
            },
            "title": "Categories",
            "type": "array"
          },
          "preserve_pinned": {
            "default": true,
            "description": "Keep POIs the user pinned and legs they have not touched, rather than discarding the whole plan.",
            "title": "Preserve Pinned",
            "type": "boolean"
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Free-text steer from the chat bar, e.g. 'prefer hot springs'.",
            "title": "Prompt"
          }
        },
        "title": "ReplanRequest",
        "type": "object"
      },
...
      "Trip": {
        "description": "A saved trip. Serialized verbatim to `trips/<slug>/trip.json`.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "edited_at": {
            "format": "date-time",
            "title": "Edited At",
            "type": "string"
          },
          "legs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TripLeg"
            },
            "title": "Legs",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "planned_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned At"
          },
          "pois": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Poi"
            },
            "title": "Pois",
            "type": "array"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "waypoints": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Waypoint"
            },
            "title": "Waypoints",
            "type": "array"
          }
        },
        "required": [
          "slug",
          "name",
          "created_at",
          "edited_at"
        ],
        "title": "Trip",
        "type": "object"
      },
...
      "TripSummary": {
        "description": "Listing entry. Avoids shipping full geometry for every trip in the index.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "edited_at": {
            "format": "date-time",
            "title": "Edited At",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "needs_replan": {
            "title": "Needs Replan",
            "type": "boolean"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "total_distance_m": {
            "title": "Total Distance M",
            "type": "number"
          }
        },
        "required": [
          "slug",
          "name",
          "created_at",
          "edited_at",
          "total_distance_m",
          "needs_replan"
        ],
        "title": "TripSummary",
        "type": "object"
      },
      "UpdateTripRequest": {
        "description": "Replace a trip's editable content.\n\nA full replacement rather than a patch: the frontend holds the authoritative trip\nstate while editing, and partial updates would need conflict rules that an\nunauthenticated prototype cannot enforce anyway.",
        "properties": {
          "legs": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/TripLeg"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legs"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "pois": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Poi"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pois"
          },
          "waypoints": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Waypoint"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waypoints"
          }
        },
        "title": "UpdateTripRequest",
        "type": "object"
      },
      "Waypoint": {
        "description": "An anchor the route must pass through.",
        "properties": {
          "coordinate": {
            "$ref": "#/components/schemas/Coordinate"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "pinned": {
            "default": true,
            "title": "Pinned",
            "type": "boolean"
          }
        },
        "required": [
          "coordinate"
        ],
        "title": "Waypoint",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "MotoRooter",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/trips": {
      "get": {
        "operationId": "list_trips_api_trips_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TripSummary"
                  },
                  "title": "Response List Trips Api Trips Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Trips",
        "tags": [
          "trips"
        ]
      },
      "post": {
        "description": "Create an empty trip.\n\nAn explicit slug is validated strictly; a derived one is slugified from the name.\nEither way it is checked before it can become a storage path.",
        "operationId": "create_trip_api_trips_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTripRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/api/trips/{slug}": {
      "delete": {
        "operationId": "delete_trip_api_trips__slug__delete",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Delete Trip",
        "tags": [
          "trips"
        ]
      },
      "get": {
        "operationId": "get_trip_api_trips__slug__get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Trip",
        "tags": [
          "trips"
        ]
      },
      "put": {
        "description": "Replace a trip's editable content.\n\n`edited_at` advances only when geometry actually changes, since it is what drives the\nreplan staleness flag \u2014 bumping it on a rename would spuriously mark discovery stale.",
        "operationId": "update_trip_api_trips__slug__put",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTripRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Update Trip",
        "tags": [
          "trips"
        ]
      }
    },
    "/api/trips/{slug}/gpx": {
      "get": {
        "description": "Returns a GPX file containing a track plus ordered waypoints, targeted at motorcycle GPS units.",
        "operationId": "export_gpx_api_trips__slug__gpx_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/gpx+xml": {}
            },
            "description": "GPX file."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Export GPX (not yet implemented)",
        "tags": [
          "trips"
        ]
      }
    },
    "/api/trips/{slug}/replan": {
      "post": {
        "description": "Runs LLM route search, POI discovery, and Places enrichment.\n\n**Streams newline-delimited JSON** (`application/x-ndjson`): one `ReplanEvent` object per line, terminated by `\\n`. Not Server-Sent Events \u2014 this is a POST with a request body, so `EventSource` cannot consume it, and hand-parsing SSE framing over `fetch` would cost the framing overhead for none of the benefit. Clients must tolerate a chunk boundary landing mid-line.\n\nExplicitly user-triggered \u2014 never fired automatically by a route edit.",
        "operationId": "replan_api_trips__slug__replan_post",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/x-ndjson": {
                "schema": {
                  "$ref": "#/components/schemas/ReplanEvent"
                }
              }
            },
            "description": "Stream of ReplanEvent objects, one per line."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not implemented yet."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Start a replan (not yet implemented)",
        "tags": [
          "trips"
        ]
      }
    }
  }
}